# ENH: added axis param for np.count_nonzero