Addresses feature request in #391 to add an axis parameter to `np.count_nonzero`.

Duplicate of #7138 after a forced push on another branch accidentally reset my origin branch for that PR to `master`, closing it automatically (I had reset hard on the #7138 branch to `master` so I could do the C refactoring @jaimefrio suggested in the spirit of #4330, whoops).