@homu 2016-02-14
:umbrella: The latest upstream changes (presumably #7246) made this pull request unmergeable. Please resolve the merge conflicts.