@gfyoung 2016-02-15
Unfortunately, I have not been able to make much progress on refactoring the code into C.  However, I think it should be good to merge as is since it performs relatively well, and the tests pass.