@gfyoung 2016-02-17
@jaimefrio : Unfortunately, I have not been able to make much progress on refactoring the code into C.  However, I think it should be good to merge as is since it performs relatively well, and the tests pass.  If someone else wants to take a look at this, that would be great as well!