@homu 2016-03-12
:umbrella: The latest upstream changes (presumably #7346) made this pull request unmergeable. Please resolve the merge conflicts.