@gfyoung 2016-03-12
Alright, this is the second time I've encountered a merge conflict on this PR.  Could someone take a look at this and try to land it?