@homu 2016-03-13
:umbrella: The latest upstream changes (presumably #7178) made this pull request unmergeable. Please resolve the merge conflicts.