@shoyer 2016-03-14
This code currently has a lot of branches. Given the apparent lack of existing test coverage, I'm not even sure all of them are being exercised. It needs to test every valid dtype, and to test invalid (e.g., `axis='foo'` or `axis=3` for a 2d array), negative and tuple axes (e.g., `axis=[0, 1]`).

I have also have some broader concerns about `count_nonzero`:

1. The behavior for non-numeric dtypes is not documented. What exactly does it mean to be "zero"? Apparently `None` and `''` are all considered zero for object arrays:
	```
	In [13]: x = np.array([None, 0, '', 12])

	In [14]: x
	Out[14]: array([None, 0, '', 12], dtype=object)

	In [15]: np.count_nonzero(x)
	Out[15]: 1
	```
    Is this behavior a bug or a feature? I don't know.

2. If `(x != 0).sum()` is faster and more explicit, what's the point of even encouraging the use of this function?