@njsmith 2016-03-14
> What exactly does it mean to be "zero"?

I haven't looked at the rest, but I can explain the name :-). In py2, "nonzero" is the official name for the property that's more commonly known as "truthiness" (see `__nonzero__`). This is confusing and in py3 they fixed this (`__nonzero__` is now `__bool__`), but this function's name remains as collateral damage.