@gfyoung 2016-03-14
@shoyer: The reason why `a != 0` isn't used by itself is because the definition of non-zero, as @njsmith enumerated, is unfortunately not as clear-cut as it seemed to be at the surface.  I too had that thought initially, and then `numpy` tests told me otherwise.