@shoyer 2016-03-14
OK, that makes sense for the definition of "nonzero". It would be good to document that :).