@gfyoung 2016-03-14
@njsmith : If we could rename it, what should it be called?  That feels like a `FutureWarning` of some kind (i.e. "WARNING: name change coming soon!"), maybe not for this PR but at least for a follow-up.