@gfyoung 2016-03-14
@shoyer: +1 for the documentation.  That has been added along with `[ci skip]` to this PR until there are other code-related changes to make OR if people deem this merge-able.