@gfyoung 2016-03-21
@shoyer, @njsmith : Is this good to merge?  I've made all of the requested changes, and I haven't seen any complaints in the past week or so.