@shoyer 2016-03-21
This still needs comprehensive test coverage, as I mentioned [in my comment above](https://github.com/numpy/numpy/pull/7177#issuecomment-196109100)