@gfyoung 2016-03-26
@shoyer, @njsmith : Added lots of new tests for `count_nonzero` that I think are fairly comprehensive.  Travis and Appveyor give the green light.  If there is nothing else, should be ready to merge.