@gfyoung 2016-04-05
@shoyer , @njsmith : I haven't seen any complaints / issue from either of you or anyone regarding this PR for almost two weeks.  Can this be merged?