@gfyoung 2016-04-17
@shoyer : I have addressed all of the important changes, except for the `_validate_axis` and conditional expressions points for reasons that I explained above.  Travis and Appveyor are both passing still.