@homu 2016-05-06
:umbrella: The latest upstream changes (presumably #7539) made this pull request unmergeable. Please resolve the merge conflicts.