@gfyoung 2016-05-06
@shoyer @njsmith @charris : Okay, this is the fourth merge conflict I've had to resolve on this PR.  Can nobody really take a look at this?