@shoyer 2016-05-07
Sorry for taking so long to get back to you.

I guess I'm off two minds for this change. It's certainly a nice feature to have. But the implementation here does feel overly complex -- you've reinvented a lot of pieces and used a lot of tricks to make this work in a performance way. The code is still far from being obvious.

Obviously my opinion is not decisive here, but I would feel more comfortable if this were done the obvious way in C, with numpy's existing iterators. As is I worry that we are adding technical debt, even if all contained in one function.