@gfyoung 2016-05-07
@shoyer : I appreciate the feedback, but why did you wait until now to give me such a critique?  Not once did I get this impression from you with all of the other comments.  As a contributor, to have been mislead like this is not very encouraging, especially since I spent a lot of time trying to optimize and take advantage of some of the other highly performing C code spread out in the codebase.