@shoyer 2016-05-07
@gfyoung  I certainly owe you an apology for not stating these concerns directly earlier. Part of this is because I was waiting for a more experienced developer to chime in, which didn't happen. I'm actually quite new to numpy and have only made small fixes to the C code myself, so I'm not actually entirely confident that what I described above is doable/reasonable. I also hoped that your code would be able to get simpler, but instead you optimized it further and added coverage for more edge cases -- perfectly reasonable in the right context, just not what I was hoping for.