@gfyoung 2016-05-22
@rgommers : Updated timing information from <a href="https://github.com/numpy/numpy/pull/7138#issuecomment-176638897">last time</a>:

1) Strings

```python -m timeit -n 100 "import numpy as np" "m=['s'] * 1000" "m[1] = ''" "np.count_nonzero(m)"```
**master**: 100 loops, best of 3, 159 usec / loop
**PR**: 100 loops, best of 3, 181 usec / loop

2) Small Integer Array

```python -m timeit -n 100 "import numpy as np" "m=np.arange(100)" "np.count_nonzero(m)"```
**master**: 100 loops, best of 3, 5.53 usec / loop
**PR**: 100 loops, best of 3, 42.3 usec / loop

3) Large Integer Array

```python -m timeit -n 100 "import numpy as np; m=np.random.randint(-1, 2, size=(10, 10, 10, 10, 10, 10));np.count_nonzero(m)"```
**master**: 100 loops, best of 3, 19 msec / loop
**PR**: 100 loops, best of 3, 18.5 msec / loop

Machine: Ubuntu 14.04, Python 2.7.11